pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tendon-hand"
version = "0.1.2"
description = "Tendon-driven dexterous hand physics/control abstraction layer"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
sim = [
    "pybullet>=3.2.5",
    "gymnasium>=1.0.0",
]
rl = [
    "stable-baselines3>=2.0",
    "torch>=2.0",
]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
all = [
    "tendon-hand[sim,rl,dev]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
# No bundled robot models — users provide their own URDF

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
